ReserveHandler_change2_seok

+++ src/kr/sparta/domain/Customer.java
@@ -41,7 +41,7 @@ public class Customer {
         this.phoneNumber = phoneNumber;
     }
 
-    public int getCash() {
+    public long getCash() {
         return cash;
     }
